Moved demo authored state to its own file
