+++ src/bar-graph/components/report-item/app.test.tsx
@@ -3,7 +3,7 @@ import { act, render } from "@testing-library/react";
 import { useInitMessage, useReportItem } from "@concord-consortium/lara-interactive-api";
 import { AppComponent } from "./app";
 import { IInteractiveState } from "../types";
-import { DemoAuthoredState } from "../../demo";
+import { DemoAuthoredState } from "../../demo-authored-state";
 
 jest.mock("@concord-consortium/lara-interactive-api", () => ({
   useReportItem: jest.fn(),
